+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Now targeting 1.21.2
 * Restore old raid behavior
   * Reverts "bugfixes" [MC-274911](https://bugs.mojang.com/browse/MC-274911) and [MC-275279](https://bugs.mojang.com/browse/MC-275279)
+  * Improve compatibility with other mods (e.g. Lithium)
 * Fix compilation
 
 # 1.0.0